requirements.txt
msoffcrypto-tool=5.0.1

